feat: widen block-id storage to sixteen bits
The kernel moved its block ids to sixteen bits some time ago; this
package still stored them one byte per block, in both its in-memory
generation buffer and its persisted chunk format. Nothing truncated at
today's registry size, but past two hundred and fifty-six entries ids
would wrap silently, modulo two hundred and fifty-six, with no error
raised in either the generation or the save path — corrupt terrain and
corrupt saved worlds, produced by adding content, which is exactly
what feature parity means doing.

The buffer and every generation pass now hold sixteen-bit elements.
Light grids and pixel buffers are deliberately untouched: those really
are one byte per voxel and have nothing to do with block ids.

The persisted format is version two, two bytes per element, packed
through an explicit view rather than by aliasing a typed array over
the buffer — the latter reads host-endian, which the kernel's own
encoder deliberately avoids, so this matches it. The element width is
taken from the kernel's exported constant rather than written as a
second literal two, so the two cannot drift apart.

Version one survives read-only, with a migration applied on load. The
save library refuses any version but the current one and has no
automatic upgrade path, so a consumer that wants old worlds to keep
opening has to do this itself: try the current schema, and on a
version mismatch decode under the old one and widen. Every write is
version two, so worlds upgrade the next time they are saved rather
than being rewritten under a bare load. Breaking existing saves was
the one outcome worse than the bug.

A test round-trips two ids that byte storage cannot hold: three hundred,
which could only ever have been stored as forty-four, and the sixteen-bit
ceiling itself. Another seals
a hand-authored version-one envelope into storage and proves it loads,
migrates, and is rewritten as version two.

The golden digests move because the hashed byte sequence doubled in
width, not because terrain changed: every other golden invariant —
bedrock floor, sea level, water, ore and tree counts — is unchanged.
Their hashing also now uses the same explicit packing as the save
format, rather than reading the buffer's host-endian memory.

+++ src/domain/chunk.ts
@@ -18,13 +18,24 @@ import { type ChunkBiomeType } from './biome.js'
 
 export type Chunk = {
   readonly coord: ChunkCoord
-  /** Flat `Uint8Array` of `CHUNK_VOLUME` block ids. See `blockIndex`. */
-  readonly blocks: Uint8Array
+  /**
+   * Flat `Uint16Array` of `CHUNK_VOLUME` block ids. See `blockIndex`.
+   *
+   * Widened from `Uint8Array` (255 ceiling) so this buffer can represent
+   * every id `@nerima-games/mc-kernel` can hand out — kernel's own block
+   * storage (`BlockState`) is two bytes per element and its `BlockId` brand
+   * accepts up to `BLOCK_ID_MAX` (`0xffff` = 65535). A `Uint16Array` element
+   * assignment still wraps modulo 65536 rather than rejecting an
+   * out-of-range write, exactly as the retired `Uint8Array` wrapped modulo
+   * 256 — the ceiling moved, the lack of a runtime guard on plain element
+   * writes did not.
+   */
+  readonly blocks: Uint16Array
   /** One biome per column, indexed `lz * CHUNK_SIZE_XZ + lx`. */
   readonly biomes: ReadonlyArray<ChunkBiomeType>
 }
 
-export const emptyBlocks = (): Uint8Array => new Uint8Array(CHUNK_VOLUME)
+export const emptyBlocks = (): Uint16Array => new Uint16Array(CHUNK_VOLUME)
 
 /**
  * Read a block, treating out-of-range as air.
@@ -35,7 +46,7 @@ export const emptyBlocks = (): Uint8Array => new Uint8Array(CHUNK_VOLUME)
  * (`packages/world/domain/terrain/ravine-carver.ts:46` among others); the
  * difference is that this version is total rather than merely unchecked.
  */
-export const readBlock = (blocks: Uint8Array, index: number): number => blocks[index] ?? AIR_BLOCK_ID
+export const readBlock = (blocks: Uint16Array, index: number): number => blocks[index] ?? AIR_BLOCK_ID
 
 export const getBlockAt = (chunk: Chunk, lx: number, y: number, lz: number): number =>
   readBlock(chunk.blocks, blockIndex(lx, y, lz))
@@ -52,7 +63,7 @@ export const getBlockAt = (chunk: Chunk, lx: number, y: number, lz: number): num
  * parameter, which is what satisfies `max-params` without touching either
  * caller.
  */
-export const setBlockAt = (blocks: Uint8Array, ...coordinate: readonly [lx: number, y: number, lz: number, block: BlockId]): void => {
+export const setBlockAt = (blocks: Uint16Array, ...coordinate: readonly [lx: number, y: number, lz: number, block: BlockId]): void => {
   const [lx, y, lz, block] = coordinate
   blocks[blockIndex(lx, y, lz)] = block
 }
